+++ src/main/java/ua/training/servlet/hospital/dao/impl/JDBCDiagnosisDao.java
@@ -65,7 +65,7 @@ public class JDBCDiagnosisDao extends JDBCGenericDao<Diagnosis> implements Diagn
     }
 
     @Override
-    public int countDiagnosesOfPatient(long patientId) {
+    public long countDiagnosesOfPatient(long patientId) {
         return count(countPatientsQuery + patientId,patientsCountLabel);
     }
 }
